evalengine: optimize NullSafeComparison
